Docker

+++ BinanceJob/BinanceJob/Models/TradeElement.cs
@@ -8,7 +8,7 @@ namespace BinanceJob.Models
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int id { get; set; }
+        public long id { get; set; }
         public string? price { get; set; }
         public string? qty { get; set; }
         public string? quoteQty { get; set; }
@@ -21,7 +21,7 @@ namespace BinanceJob.Models
         public string? namePart { get; set; }
         public string? checkColumn { get; set; }
 
-        public TradeElement(int id, string? price, string? qty, string? quoteQty, long time, bool isBuyerMaker, bool isBestMatch, string? namePart)
+        public TradeElement(long id, string? price, string? qty, string? quoteQty, long time, bool isBuyerMaker, bool isBestMatch, string? namePart)
         {
             this.id = id;
             this.price = price;
